backend/requirements.txt
flask==1.1.2
flask_sqlalchemy==2.4.4
SQLAlchemy<1.4
flask_script==2.0.6
flask_migrate==2.7.0
flask_login==0.5.0
mysqlclient==2.0.3
mysql-connector-python==8.0.13
requests==2.25.1
cryptography
flask-cors